Added documentation comments to Kademlia classes. Moved classes into separate files. Organised Kadelia package.
