+++ src/main/kotlin/kademlia/KademliaQueueMessage.kt
@@ -1,0 +1,8 @@
+package kademlia
+
+/**
+ * Created by mihael
+ * on 10/12/2021 at 20:27
+ * using IntelliJ IDEA
+ */
+class QueueMessage(val ip: String, val port: Int, val data: ByteArray)
