--- src/main/kotlin/network/MessageEndpoint.kt
@@ -1,5 +1,0 @@
-package network
-
-import data.network.Endpoint
-
-annotation class MessageEndpoint(val endpoint: Endpoint)
